Crate google-adsensehost4d1

This documentation was generated from AdSense Host crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

accounts (adclients.get, adclients.list, adunits.delete, adunits.get, adunits.get ad code, adunits.insert, adunits.list, adunits.patch, adunits.update, get, list and reports.generate)

adclients (get and list)

associationsessions (start and verify)

customchannels (delete, get, insert, list, patch and update)

reports (generate)

urlchannels (delete, insert and list)

Everything else about the AdSense Host v4d1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.accounts().adunits_get(...).doit()
let r = hub.accounts().get(...).doit()
let r = hub.accounts().adunits_list(...).doit()
let r = hub.accounts().adunits_get_ad_code(...).doit()
let r = hub.accounts().reports_generate(...).doit()
let r = hub.accounts().adunits_delete(...).doit()
let r = hub.accounts().adunits_update(...).doit()
let r = hub.accounts().adunits_patch(...).doit()
let r = hub.accounts().adunits_insert(...).doit()
let r = hub.accounts().list(...).doit()
let r = hub.accounts().adclients_list(...).doit()
let r = hub.accounts().adclients_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-adsensehost4d1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-adsensehost4d1" as adsensehost4d1;
use adsensehost4d1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsensehost4d1::AdSenseHost;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSenseHost::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().reports_generate("accountId", "startDate", "endDate")
             .start_index(92)
             .add_sort("nonumy")
             .add_metric("dolores")
             .max_results(40)
             .locale("sadipscing")
             .add_filter("aliquyam")
             .add_dimension("ea")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

Account
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits get accounts (none)

get accounts (response)

adunits list accounts (none)

adunits get ad code accounts (none)

reports generate accounts (none)

adunits delete accounts (none)

adunits update accounts (none)

adunits patch accounts (none)

adunits insert accounts (none)

list accounts (none)

adclients list accounts (none)

adclients get accounts (none)

AccountAdclientGetMethodBuilder
Get information about one of the ad clients in the specified publisher's AdSense account.

AccountAdclientListMethodBuilder
List all hosted ad clients in the specified hosted account.

AccountAdunitDeleteMethodBuilder
Delete the specified ad unit from the specified publisher AdSense account.

AccountAdunitGetAdCodeMethodBuilder
Get ad code for the specified ad unit, attaching the specified host custom channels.

AccountAdunitGetMethodBuilder
Get the specified host ad unit in this AdSense account.

AccountAdunitInsertMethodBuilder
Insert the supplied ad unit into the specified publisher AdSense account.

AccountAdunitListMethodBuilder
List all ad units in the specified publisher's AdSense account.

AccountAdunitPatchMethodBuilder
Update the supplied ad unit in the specified publisher AdSense account. This method supports patch semantics.

AccountAdunitUpdateMethodBuilder
Update the supplied ad unit in the specified publisher AdSense account.

AccountGetMethodBuilder
Get information about the selected associated AdSense account.

AccountListMethodBuilder
List hosted accounts associated with this AdSense account by ad client id.

AccountMethodsBuilder
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdSenseHost hub.

AccountReportGenerateMethodBuilder
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

Accounts
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list accounts (response)

AdClient
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get adclients (response)

adclients get accounts (response)

AdClients
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list adclients (response)

adclients list accounts (response)

AdCode
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits get ad code accounts (response)

AdSenseHost
Central instance to access all AdSenseHost related resource activities

AdStyle
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

AdStyleColors
The colors included in the style. These are represented as six hexadecimal characters, similar to HTML color codes, but without the leading hash.

This type is not used in any activity, and only used as part of another schema.

AdStyleFont
The font which is included in the style.

This type is not used in any activity, and only used as part of another schema.

AdUnit
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits get accounts (response)

adunits patch accounts (request|response)

adunits update accounts (request|response)

adunits delete accounts (response)

adunits insert accounts (request|response)

AdUnitContentAdsSettings
Settings specific to content ads (AFC) and highend mobile content ads (AFMC).

This type is not used in any activity, and only used as part of another schema.

AdUnitContentAdsSettingsBackupOption
The backup option to be used in instances where no ad is available.

This type is not used in any activity, and only used as part of another schema.

AdUnitMobileContentAdsSettings
Settings specific to WAP mobile content ads (AFMC).

This type is not used in any activity, and only used as part of another schema.

AdUnits
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits list accounts (response)

AdclientGetMethodBuilder
Get information about one of the ad clients in the Host AdSense account.

AdclientListMethodBuilder
List all host ad clients in this AdSense account.

AdclientMethodsBuilder
A builder providing access to all methods supported on adclient resources. It is not used directly, but through the AdSenseHost hub.

AssociationSession
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

verify associationsessions (response)

start associationsessions (response)

AssociationsessionMethodsBuilder
A builder providing access to all methods supported on associationsession resources. It is not used directly, but through the AdSenseHost hub.

AssociationsessionStartMethodBuilder
Create an association session for initiating an association with an AdSense user.

AssociationsessionVerifyMethodBuilder
Verify an association session after the association callback returns from AdSense signup.

CustomChannel
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get customchannels (response)

update customchannels (request|response)

delete customchannels (response)

insert customchannels (request|response)

patch customchannels (request|response)

CustomChannels
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list customchannels (response)

CustomchannelDeleteMethodBuilder
Delete a specific custom channel from the host AdSense account.

CustomchannelGetMethodBuilder
Get a specific custom channel from the host AdSense account.

CustomchannelInsertMethodBuilder
Add a new custom channel to the host AdSense account.

CustomchannelListMethodBuilder
List all host custom channels in this AdSense account.

CustomchannelMethodsBuilder
A builder providing access to all methods supported on customchannel resources. It is not used directly, but through the AdSenseHost hub.

CustomchannelPatchMethodBuilder
Update a custom channel in the host AdSense account. This method supports patch semantics.

CustomchannelUpdateMethodBuilder
Update a custom channel in the host AdSense account.

Report
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reports generate accounts (response)

generate reports (response)

ReportGenerateMethodBuilder
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

ReportHeaders
The header information of the columns requested in the report. This is a list of headers; one for each dimension in the request, followed by one for each metric in the request.

This type is not used in any activity, and only used as part of another schema.

ReportMethodsBuilder
A builder providing access to all methods supported on report resources. It is not used directly, but through the AdSenseHost hub.

UrlChannel
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete urlchannels (response)

insert urlchannels (request|response)

UrlChannels
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list urlchannels (response)

UrlchannelDeleteMethodBuilder
Delete a URL channel from the host AdSense account.

UrlchannelInsertMethodBuilder
Add a new URL channel to the host AdSense account.

UrlchannelListMethodBuilder
List all host URL channels in the host AdSense account.

UrlchannelMethodsBuilder
A builder providing access to all methods supported on urlchannel resources. It is not used directly, but through the AdSenseHost hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.