Module google_adsensehost4d1::api

Structs

Account
There is no detailed description.

AccountAdclientGetCall
Get information about one of the ad clients in the specified publisher's AdSense account.

AccountAdclientListCall
List all hosted ad clients in the specified hosted account.

AccountAdunitDeleteCall
Delete the specified ad unit from the specified publisher AdSense account.

AccountAdunitGetAdCodeCall
Get ad code for the specified ad unit, attaching the specified host custom channels.

AccountAdunitGetCall
Get the specified host ad unit in this AdSense account.

AccountAdunitInsertCall
Insert the supplied ad unit into the specified publisher AdSense account.

AccountAdunitListCall
List all ad units in the specified publisher's AdSense account.

AccountAdunitPatchCall
Update the supplied ad unit in the specified publisher AdSense account. This method supports patch semantics.

AccountAdunitUpdateCall
Update the supplied ad unit in the specified publisher AdSense account.

AccountGetCall
Get information about the selected associated AdSense account.

AccountListCall
List hosted accounts associated with this AdSense account by ad client id.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdSenseHost hub.

AccountReportGenerateCall
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

Accounts
There is no detailed description.

AdClient
There is no detailed description.

AdClients
There is no detailed description.

AdCode
There is no detailed description.

AdSenseHost
Central instance to access all AdSenseHost related resource activities

AdStyle
There is no detailed description.

AdStyleColors
The colors included in the style. These are represented as six hexadecimal characters, similar to HTML color codes, but without the leading hash.

AdStyleFont
The font which is included in the style.

AdUnit
There is no detailed description.

AdUnitContentAdsSettings
Settings specific to content ads (AFC) and highend mobile content ads (AFMC - deprecated).

AdUnitContentAdsSettingsBackupOption
The backup option to be used in instances where no ad is available.

AdUnitMobileContentAdsSettings
Settings specific to WAP mobile content ads (AFMC - deprecated).

AdUnits
There is no detailed description.

AdclientGetCall
Get information about one of the ad clients in the Host AdSense account.

AdclientListCall
List all host ad clients in this AdSense account.

AdclientMethods
A builder providing access to all methods supported on adclient resources. It is not used directly, but through the AdSenseHost hub.

AssociationSession
There is no detailed description.

AssociationsessionMethods
A builder providing access to all methods supported on associationsession resources. It is not used directly, but through the AdSenseHost hub.

AssociationsessionStartCall
Create an association session for initiating an association with an AdSense user.

AssociationsessionVerifyCall
Verify an association session after the association callback returns from AdSense signup.

CustomChannel
There is no detailed description.

CustomChannels
There is no detailed description.

CustomchannelDeleteCall
Delete a specific custom channel from the host AdSense account.

CustomchannelGetCall
Get a specific custom channel from the host AdSense account.

CustomchannelInsertCall
Add a new custom channel to the host AdSense account.

CustomchannelListCall
List all host custom channels in this AdSense account.

CustomchannelMethods
A builder providing access to all methods supported on customchannel resources. It is not used directly, but through the AdSenseHost hub.

CustomchannelPatchCall
Update a custom channel in the host AdSense account. This method supports patch semantics.

CustomchannelUpdateCall
Update a custom channel in the host AdSense account.

Report
There is no detailed description.

ReportGenerateCall
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

ReportHeaders
The header information of the columns requested in the report. This is a list of headers; one for each dimension in the request, followed by one for each metric in the request.

ReportMethods
A builder providing access to all methods supported on report resources. It is not used directly, but through the AdSenseHost hub.

UrlChannel
There is no detailed description.

UrlChannels
There is no detailed description.

UrlchannelDeleteCall
Delete a URL channel from the host AdSense account.

UrlchannelInsertCall
Add a new URL channel to the host AdSense account.

UrlchannelListCall
List all host URL channels in the host AdSense account.

UrlchannelMethods
A builder providing access to all methods supported on urlchannel resources. It is not used directly, but through the AdSenseHost hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.