Struct google_adsensehost4d1::AccountMethods

pub struct AccountMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on account resources. It is not used directly, but through the AdSenseHost hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adsensehost4d1 as adsensehost4d1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsensehost4d1::AdSenseHost;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSenseHost::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `adclients_get(...)`, `adclients_list(...)`, `adunits_delete(...)`, `adunits_get(...)`, `adunits_get_ad_code(...)`, `adunits_insert(...)`, `adunits_list(...)`, `adunits_patch(...)`, `adunits_update(...)`, `get(...)`, `list(...)` and `reports_generate(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, A> AccountMethods<'a, C, A>

fn adunits_list(&self, account_id: &str, ad_client_id: &str) -> AccountAdunitListCall<'a, C, A>

Create a builder to help you perform the following task:

List all ad units in the specified publisher's AdSense account.

Arguments

accountId - Account which contains the ad client.

adClientId - Ad client for which to list ad units.

fn list(&self, filter_ad_client_id: &Vec<String>) -> AccountListCall<'a, C, A>

Create a builder to help you perform the following task:

List hosted accounts associated with this AdSense account by ad client id.

Arguments

filterAdClientId - Ad clients to list accounts for.

fn adunits_get_ad_code(&self, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitGetAdCodeCall<'a, C, A>

Create a builder to help you perform the following task:

Get ad code for the specified ad unit, attaching the specified host custom channels.

Arguments

accountId - Account which contains the ad client.

adClientId - Ad client with contains the ad unit.

adUnitId - Ad unit to get the code for.

fn reports_generate(&self, account_id: &str, start_date: &str, end_date: &str) -> AccountReportGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

Arguments

accountId - Hosted account upon which to report.

startDate - Start of the date range to report on in "YYYY-MM-DD" format, inclusive.

endDate - End of the date range to report on in "YYYY-MM-DD" format, inclusive.

fn adunits_delete(&self, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete the specified ad unit from the specified publisher AdSense account.

Arguments

accountId - Account which contains the ad unit.

adClientId - Ad client for which to get ad unit.

adUnitId - Ad unit to delete.

fn adclients_list(&self, account_id: &str) -> AccountAdclientListCall<'a, C, A>

Create a builder to help you perform the following task:

List all hosted ad clients in the specified hosted account.

Arguments

accountId - Account for which to list ad clients.

fn adunits_patch(&self, request: AdUnit, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update the supplied ad unit in the specified publisher AdSense account. This method supports patch semantics.

Arguments

request - No description provided.

accountId - Account which contains the ad client.

adClientId - Ad client which contains the ad unit.

adUnitId - Ad unit to get.

fn adclients_get(&self, account_id: &str, ad_client_id: &str) -> AccountAdclientGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get information about one of the ad clients in the specified publisher's AdSense account.

Arguments

accountId - Account which contains the ad client.

adClientId - Ad client to get.

fn get(&self, account_id: &str) -> AccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get information about the selected associated AdSense account.

Arguments

accountId - Account to get information about.

fn adunits_insert(&self, request: AdUnit, account_id: &str, ad_client_id: &str) -> AccountAdunitInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Insert the supplied ad unit into the specified publisher AdSense account.

Arguments

request - No description provided.

accountId - Account which will contain the ad unit.

adClientId - Ad client into which to insert the ad unit.

fn adunits_get(&self, account_id: &str, ad_client_id: &str, ad_unit_id: &str) -> AccountAdunitGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get the specified host ad unit in this AdSense account.

Arguments

accountId - Account which contains the ad unit.

adClientId - Ad client for which to get ad unit.

adUnitId - Ad unit to get.

fn adunits_update(&self, request: AdUnit, account_id: &str, ad_client_id: &str) -> AccountAdunitUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update the supplied ad unit in the specified publisher AdSense account.

Arguments

request - No description provided.

accountId - Account which contains the ad client.

adClientId - Ad client which contains the ad unit.