Struct google_adsensehost4d1::CustomchannelMethods

pub struct CustomchannelMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on customchannel resources. It is not used directly, but through the AdSenseHost hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adsensehost4d1 as adsensehost4d1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsensehost4d1::AdSenseHost;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSenseHost::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.customchannels();

Methods

impl<'a, C, A> CustomchannelMethods<'a, C, A>

fn get(&self, ad_client_id: &str, custom_channel_id: &str) -> CustomchannelGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a specific custom channel from the host AdSense account.

Arguments

adClientId - Ad client from which to get the custom channel.

customChannelId - Custom channel to get.

fn delete(&self, ad_client_id: &str, custom_channel_id: &str) -> CustomchannelDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a specific custom channel from the host AdSense account.

Arguments

adClientId - Ad client from which to delete the custom channel.

customChannelId - Custom channel to delete.

fn update(&self, request: CustomChannel, ad_client_id: &str) -> CustomchannelUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update a custom channel in the host AdSense account.

Arguments

request - No description provided.

adClientId - Ad client in which the custom channel will be updated.

fn list(&self, ad_client_id: &str) -> CustomchannelListCall<'a, C, A>

Create a builder to help you perform the following task:

List all host custom channels in this AdSense account.

Arguments

adClientId - Ad client for which to list custom channels.

fn patch(&self, request: CustomChannel, ad_client_id: &str, custom_channel_id: &str) -> CustomchannelPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update a custom channel in the host AdSense account. This method supports patch semantics.

Arguments

request - No description provided.

adClientId - Ad client in which the custom channel will be updated.

customChannelId - Custom channel to get.

fn insert(&self, request: CustomChannel, ad_client_id: &str) -> CustomchannelInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Add a new custom channel to the host AdSense account.

Arguments

request - No description provided.

adClientId - Ad client to which the new custom channel will be added.