Struct google_adsensehost4d1::ReportMethods

pub struct ReportMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on report resources. It is not used directly, but through the AdSenseHost hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adsensehost4d1 as adsensehost4d1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsensehost4d1::AdSenseHost;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSenseHost::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `generate(...)`
// to build up your call.
let rb = hub.reports();

Methods

impl<'a, C, A> ReportMethods<'a, C, A>

fn generate(&self, start_date: &str, end_date: &str) -> ReportGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

Arguments

startDate - Start of the date range to report on in "YYYY-MM-DD" format, inclusive.

endDate - End of the date range to report on in "YYYY-MM-DD" format, inclusive.